% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMM.r
\name{glmm}
\alias{glmm}
\title{Calculate linear mixed effects coefficients and covariance matrix (identity link)
Code can be found in: http://sia.webpopix.org/EMlme.html}
\usage{
glmm(y, id, fixed, mixed = fixed, niter = 50)
}
\arguments{
\item{y}{a vector of the outcome variable}

\item{id}{a vector of the individual level id variable}

\item{fixed}{a matrix of fixed effects (intercept term optional)}

\item{mixed}{a matrix of random effects (intercept term optional)}

\item{niter}{numeric for number of EM iterations}
}
\value{
a list of fixed effects coefficients, random effects cov. matrix,
and model selection metrics (logLik, AIC, BIC).
}
\description{
This code calculates the coefficients of the fixed effects and
covariance matrix of the random effects specified in the model.
}
\examples{

X <- cbind(seq(1,100,1), seq(1,200,2), seq(1,500,5))
glmm(X[,3], X[,1], X[,2], X[,2])

}
