% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRLS.R
\name{irls}
\alias{irls}
\title{Iteratively Reweighted Least Squares}
\usage{
irls(data, yi, xi, niter = 200)
}
\arguments{
\item{data}{a data frame}

\item{yi}{either a character string or the column number of the outcome variable in the data frame.}

\item{xi}{a vector of either character strings or the column numbers of the covariates.}

\item{niter}{numeric for number of iterations (optional argument)}
}
\value{
a matrix of the parameter estimates (includes intercept), standard errors, 95\% confidence intervals, t statistics, and p-values
}
\description{
This function estimates parameters using the iteratively reweighted least squares method.
}
\examples{

est <- irls(data=data, yi="y", xi=c("x1","x2","x3"))

est <- irls(data=data, yi=1, xi=c(1:3))

}
